Tuning parameter selection

Tuning parameter selection

tune_params_ml(gam_param, lasso_param, randomforest_param, knn_param,
  svm_param, nn_param, bart_param, folds, xnam, tao, data, weighting)

Arguments

gam_param
a vector containing degree of freedom 3 and 4

lasso_param
a grid of values for the shrinkage term

randomforest_param
a two column matrix: first column denotes the num_trees parameter and the second column denotes the mtry parameter.

knn_param
a grid of positive integers values

svm_param
a three column matrix: first column denotes the cost parameter, second column the gamma and third column the kernel. kernel=1 denotes "radial" and kernel=2 denotes "linear".

nn_param
a grid of positive integers values for the neurons

bart_param
a three column matrix: first column denotes the num_tree parameter, second column the k parameter and third column the q parameter.

folds
number of folds

xnam
vector with the names of the covariates to be included in the model

tao
evaluation time point of interest

data
data set that contains at least id, E , ttilde, delta, wts and covariates

weighting
Procedure to compute the inverse probability of censoring weights. Weighting="CoxPH" and weighting="CoxBoost" model the censoring by the Cox model and CoxBoost model respectively.

Value

a list with the tune parameters selected using the IPCW AUC loss function.